refactor(vue): update naming and file extensions for stories
