+++ packages/vue/src/tabs/stories/basic.stories.vue
