+++ packages/vue/src/tabs/stories/withDefaultValue.stories.vue
